requirements/base.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coverage==4.5.3
coveralls==1.8.1
cryptography==2.7
Django==2.2.3
django-filter==2.1.0
django-pytest==0.2.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
idna==2.7
importlib-metadata==0.18
jsonschema==2.6.0
Markdown==3.1.1
more-itertools==7.1.0
packaging==19.0
paramiko==2.6.0
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.0
pytest==5.0.1
pytest-cov==2.7.1
pytest-sugar==0.9.2
python-decouple==3.1
python-dotenv==0.10.3
pytz==2019.1
PyYAML==3.13
requests==2.20.1
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
toml==0.10.0
tox==3.13.2
urllib3==1.24.3
virtualenv==16.6.2
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.5.2
